Fix implementation and design of Location and IncidentComponent
Remove unnecessary methods in Location, and improve design of
IncidentComponent. Everything still works the same in terms of public
interface methods.

Fix bug in IncidentComponent, which checks for bad input data in
constructor, and replaces it with a default value, as specified in
DatabaseColumn Enum.

Update other classes to use the new Location class methods.